refactor: use specific modules for gui and cli
