+++ src-tauri/src/cli.rs
@@ -1,0 +1,3 @@
+pub fn run() {
+    println!("--> [cli] run");
+}
